requirements.txt
langchain
langchain_community
python-dotenv
streamlit
google-generativeai
langchain_google_genai
pillow
validators
PyPDF2
faiss-cpu
beautifulsoup4
joblib